SLX Simulink Circuit
WIP. Some discrete electronic components do not have a 1-to-1 correspondence so there will be abstracted components (i.e. MOC3021) that will be traced as two thyristors, etc..  Additional modifications will be needed using converters between the Simulink and Simscape parts of this schematic to represent the logic of the ESP32 C6.